Added dataTable js-files for better table handling
dataTable is a jquery-Plugin developed from MIT. Adding the class-type
datatable to tables makes them, among others features, sortable.
Does not work for the userDetailCourses-table and the tables in tabs, like
in courseDetails. Needs to be fixed.
dataTable can be a big plugin. Not all extensions are implemented now.
Design is not modified yet.