Paginate buttons are nearly perfect.
Size numberas have enough space between them. The datatables do not
collapse at the right time. No idea how to fix this. In the end a lot of
css-file can be deleted. All files which are noch included in the
index.php
Please enter the commit message for your changes. Lines starting